+++ CHANGELOG.md
@@ -14,11 +14,12 @@
 ### 📚 Documentation
 
 * [#57](https://github.com/antoinezanardi/werewolves-assistant-api/issues/57) - Fix doc for `game.won`.
+* [#62](https://github.com/antoinezanardi/werewolves-assistant-api/issues/62) - Icon for witch 🪄 and raven 🪶 changed.
 
 ### 📦 Packages
 
 * `@sentry/node` updated to version `5.27.4`.
-* `mongoose` updated to version `5.10.14`.
+* `mongoose` updated to version `5.10.15`.
 
 ---